# exec: template comparisons for ordered sync