--- a/pkg/sql/exec/.gitignore
+++ b/pkg/sql/exec/.gitignore
@@ -14,3 +14,4 @@ selection_ops.eg.go
 sort.eg.go
 sum_agg.eg.go
 tuples_differ.eg.go
+vec_comparators.eg.go