--- a/pkg/sql/exec/vec_comparators.go
+++ b/pkg/sql/exec/vec_comparators.go
@@ -1,0 +1,30 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package exec
+
+import "github.com/cockroachdb/cockroach/pkg/sql/exec/coldata"
+
+// vecComparator is a helper for the ordered synchronizer. It stores multiple
+// column vectors of a single type and facilitates comparing values between
+// them. The implementations are templated by vec_comparators_gen.go.
+type vecComparator interface {
+	// compare compares values from two vectors. vecIdx is the index of the vector
+	// and valIdx is the index of the value in that vector to compare. Returns -1,
+	// 0, or 1.
+	compare(vecIdx1, vecIdx2 int, valIdx1, valIdx2 uint16) int
+
+	// setVec updates the vector at idx.
+	setVec(idx int, vec coldata.Vec)
+}