The ordered synchronizer now supports all types rather than just int64.
To implement this I introduced a vecComparator interface which compares
values across multiple same-typed vectors. The synchronizer creates one
of these for each ordering column.

Introducing the comparators provided a significant speed-up since it
prevents needing to repeatedly call `Vec.Int64()`.
```
name                   old time/op   new time/op   delta
OrderedSynchronizer-4   94.7µs ± 1%   61.5µs ± 1%  -35.09%  (p=0.008 n=5+5)

name                   old speed     new speed     delta
OrderedSynchronizer-4  259MB/s ± 1%  400MB/s ± 1%  +54.05%  (p=0.008 n=5+5)
```

Refers #37304

Release note: None